+++ platform/linux/bootstrap.sh
@@ -37,6 +37,7 @@ fi
 "$ROOT/platform/linux/install-oh-my-posh.sh"
 "$ROOT/platform/linux/install-docker.sh"
 "$ROOT/platform/linux/install-vscode.sh" || true
+"$ROOT/platform/linux/install-gh.sh" || true
 "$ROOT/platform/linux/install-librewolf.sh" || true
 "$ROOT/platform/linux/install-solaar.sh" || true
 "$ROOT/platform/linux/install-wireshark.sh" || true